+++ base/gui/searchAllDialog.py
@@ -1,109 +1,74 @@
 import re
 from dataclasses import dataclass, field
-from typing import Optional, Sequence, Union
+from typing import Callable, Iterator, Optional, Sequence
 
-from PyQt5.QtCore import QEventLoop, pyqtSignal, QTimer
 from PyQt5.QtGui import QKeySequence
-from PyQt5.QtWidgets import QDialog, QWidget, QApplication
+from PyQt5.QtWidgets import QWidget
+from recordclass import as_dataclass
 
-from cat.GUI import CORNERS, PythonGUI
-from cat.GUI.components.Widgets import HTMLDelegate
-from cat.GUI.components.codeEditor import SearchOptions, SearchMode
-from cat.GUI.components.treeBuilders import DataTreeBuilder
-from cat.GUI.framelessWindow.catFramelessWindowMixin import CatFramelessWindowMixin
-from cat.GUI.utilities import connectOnlyOnce
-from gui.icons import icons
-from cat.utils import escapeForXml
-from cat.utils.collections_ import OrderedMultiDict
-from cat.utils.profiling import TimedMethod
-from base.model.project.project import Root
+from base.gui.onProjectFilesDialogBase import OnProjectFilesDialogBase
+from base.model.pathUtils import FilePathTpl, ZipFilePool, dirFromFilePath, fileNameFromFilePath, loadTextFile
 from base.model.session import getSession
-from basePlugins.projectFiles import FilesIndex
-from gui.datapackEditorGUI import ContextMenuEntries, makeTextSearcher
-from base.model.pathUtils import FilePath, ZipFilePool, loadTextFile
 from base.model.utils import Position, Span
+from cat.GUI import CORNERS
+from cat.GUI.components.codeEditor import IndexSpan, SearchMode, SearchOptions
+from cat.GUI.components.treeBuilders import DataTreeBuilder
+from cat.utils import escapeForXml, override
+from cat.utils.collections_ import OrderedMultiDict
+from gui.datapackEditorGUI import ContextMenuEntries, DatapackEditorGUI, makeTextSearcher
+from gui.icons import icons
 
 
-@dataclass(unsafe_hash=True)
+@as_dataclass(hashable=True)
 class Occurrence:
-	file: FilePath
+	file: FilePathTpl
 	span: Span
 	line: str
 
 
 @dataclass
 class SearchResult:
-	filesToSearch: list[FilePath] = field(default_factory=list)
-	occurrences: OrderedMultiDict[FilePath, Occurrence] = field(default_factory=OrderedMultiDict)
-	filesSearched: int = 0
+	occurrences: OrderedMultiDict[FilePathTpl, Occurrence] = field(default_factory=OrderedMultiDict)
 	error: Optional[Exception] = None
 
 
-class SearchAllDialog(CatFramelessWindowMixin, QDialog):
+class SearchAllDialog(OnProjectFilesDialogBase):
 
 	def __init__(self, parent: Optional[QWidget] = None):
-		super().__init__(GUICls=PythonGUI, parent=parent)
+		super().__init__(GUICls=DatapackEditorGUI, parent=parent)
 
-		self._includedRoots: list[Root] = []
 		self.searchExpr: str = ''
 		self.searchOptions: SearchOptions = SearchOptions(
 			searchMode=SearchMode.Normal,
 			isCaseSensitive=False,
 			isMultiLine=False,
 		)
-
-		self._fileTypesStr: str = ""
-		self._fileTypes: tuple[str, ...] = ()
 		self._searchResult: SearchResult = SearchResult()
-		self.htmlDelegate = HTMLDelegate()
 
 		self.setWindowTitle('Search')
 
-	progressSignal = pyqtSignal(int)
-
+	@override
 	def resetUserInterface(self):
-		allFilePaths = self.filePathsToSearch
-		self._searchResult = SearchResult(allFilePaths)
+		super().resetUserInterface()
+		self._searchResult = SearchResult()
 
-	def OnSidebarGUI(self, gui: PythonGUI):
-		with gui.vLayout1C(preventVStretch=False):
-			self._fileTypesStr = gui.codeField(self._fileTypesStr, label="File Extensions:", isMultiline=False, placeholderText="All Extensions")
-			self._fileTypes = tuple(ext.strip() for ext in self._fileTypesStr.split(',') if ext.strip())
-
-		gui.vSeparator()
-		includedRoots = []
-		with gui.vLayout(preventVStretch=True, verticalSpacing=0):
-			for root in getSession().project.roots:
-				if gui.checkboxLeft(None, root.name):
-					includedRoots.append(root)
-			gui.vSeparator()
-			for root in getSession().project.deepDependencies:
-				if gui.checkboxLeft(None, root.name):
-					includedRoots.append(root)
-		self._includedRoots = includedRoots
-
-	def OnGUI(self, gui: PythonGUI):
+	@override
+	def optionsGUI(self, gui: DatapackEditorGUI):
 		with gui.vLayout(preventVStretch=False):
-			with gui.vLayout(preventVStretch=False):
-				with gui.hLayout(horizontalSpacing=0):
-					self.searchExpr = gui.codeField(self.searchExpr, isMultiline=False, roundedCorners=CORNERS.NONE)
-					if gui.toolButton(icon=icons.search, overlap=(1, 0), roundedCorners=(False, True, False, True), default=True, windowShortcut=QKeySequence("Return")):
-						self.resetUserInterface()
-						QTimer.singleShot(1, self.search)
-				if self.searchOptions.searchMode == SearchMode.RegEx:
-					try:
-						re.compile(self.searchExpr)
-					except Exception as e:
-						gui.helpBox(str(e),  'error', hasLabel=False)
+			with gui.hLayout(horizontalSpacing=0):
+				self.searchExpr = gui.codeField(self.searchExpr, isMultiline=False, roundedCorners=CORNERS.NONE)
+				if gui.toolButton(icon=icons.search, overlap=(1, 0), roundedCorners=(False, True, False, True), default=True, windowShortcut=QKeySequence("Return")):
+					self.run()
+			if self.searchOptions.searchMode == SearchMode.RegEx:
+				try:
+					re.compile(self.searchExpr)
+				except Exception as e:
+					gui.helpBox(str(e), 'error', hasLabel=False)
 
-			self._searchOptionsGUI(gui)
-			gui.progressBar(self.progressSignal, min=0, max=len(self._searchResult.filesToSearch), value=self._searchResult.filesSearched, format='', textVisible=True)
-			resultsGUI = gui.subGUI(PythonGUI, self._resultsGUI1, suppressRedrawLogging=False)
-			connectOnlyOnce(self, self.progressSignal, lambda i: resultsGUI.redrawGUI(), 'resultsGUI')
-			resultsGUI.redrawGUI()
-			self._resultsGUI2(gui)
+		self._searchOptionsGUI(gui)
+		self.addProgressBar(gui)
 
-	def _searchOptionsGUI(self, gui: PythonGUI):
+	def _searchOptionsGUI(self, gui: DatapackEditorGUI):
 		so = self.searchOptions
 		# ============ Search Options: ============
 		with gui.hLayout(preventHStretch=True):
@@ -115,33 +80,37 @@ class SearchAllDialog(CatFramelessWindowMixin, QDialog):
 			so.isCaseSensitive = gui.toggleLeft(so.isCaseSensitive, 'Case sensitive')
 			so.isMultiLine = gui.toggleLeft(so.isMultiLine, 'Multiline', enabled=so.searchMode == SearchMode.RegEx)
 
-	def _resultsGUI1(self, gui: PythonGUI) -> None:
+	@override
+	def resultsSummaryGUI(self, gui: DatapackEditorGUI) -> None:
 		if self._searchResult.error is not None:
 			gui.helpBox(f'error during search: {self._searchResult.error}', style='error')
 		else:
-			gui.label(f'found {len(self._searchResult.occurrences)} occurrences in {len(self._searchResult.occurrences.uniqueKeys())} files ({self._searchResult.filesSearched} files searched total): (double-click to open)')
+			gui.label(f'found {len(self._searchResult.occurrences)} occurrences in {len(self._searchResult.occurrences.uniqueKeys())} files ({self.processedFilesCount} files searched total): (double-click to open)')
 
-	def _resultsGUI2(self, gui: PythonGUI) -> None:
-		def labelMaker(x:  Union[SearchResult, FilePath, Occurrence], i: int) -> str:
+	@override
+	def resultsGUI(self, gui: DatapackEditorGUI) -> None:
+		def labelMaker(x: SearchResult | FilePathTpl | Occurrence, column: int) -> str:
 			if isinstance(x, Occurrence):
-				return x.line
+				return x.line if column == 0 else str(x.span.start)
+			elif isinstance(x, SearchResult):
+				return '<root>'
 			else:
-				countInFile = len(self._searchResult.occurrences.getall(x))
-				if isinstance(x, tuple):
-					filename = x[1].rpartition('/')[2]
-					return f'{filename} - ({countInFile}) - "{str(x[0])}"'
-				elif isinstance(x, str):
-					filename = x.rpartition('/')[2]
-					return f'{filename} - (countInFile) - "{str(x[0])}"'
-			return '<root>'
+				if column == 0:
+					fileName = fileNameFromFilePath(x[1])
+					projectName = fileNameFromFilePath(x[0])
+					pathInProj = dirFromFilePath(x[1]).removeprefix('/')
+					countInFile = len(self._searchResult.occurrences.getall(x))
+					return f'{fileName} - ({countInFile}) - "{projectName}/{pathInProj}"'
+				else:
+					return ''
 
-		def openDocument(x: Union[FilePath, Occurrence], *, s=self):
+		def openDocument(x: SearchResult | FilePathTpl | Occurrence) -> None:
 			if isinstance(x, Occurrence):
 				getSession().tryOpenOrSelectDocument(x.file, x.span)
 			else:
 				getSession().tryOpenOrSelectDocument(x)
 
-		def onContextMenu(x: Union[SearchResult, FilePath, Occurrence], column: int, *, s=self):
+		def onContextMenu(x: SearchResult | FilePathTpl | Occurrence, column: int) -> None:
 			if isinstance(x, Occurrence):
 				with gui.popupMenu(atMousePosition=True) as menu:
 					menu.addItems(ContextMenuEntries.fileItems(x.file, getSession().tryOpenOrSelectDocument))
@@ -149,81 +118,85 @@ class SearchAllDialog(CatFramelessWindowMixin, QDialog):
 				with gui.popupMenu(atMousePosition=True) as menu:
 					menu.addItems(ContextMenuEntries.fileItems(x, getSession().tryOpenOrSelectDocument))
 
-		def childrenMaker(x: Union[SearchResult, FilePath, Occurrence], *, s=self) -> Sequence:
-			if isinstance(x, SearchResult):
-				return list(x.occurrences.uniqueKeys())
-				# return [(fp, x.occurrences.getall(fp)) for fp in x.occurrences.uniqueKeys()]
-			elif isinstance(x, Occurrence):
+		def childrenMaker(x: SearchResult | FilePathTpl | Occurrence) -> Sequence[FilePathTpl | Occurrence]:
+			if isinstance(x, Occurrence):
 				return tuple()
+			elif isinstance(x, SearchResult):
+				return list(x.occurrences.uniqueKeys())
 			else:
 				return self._searchResult.occurrences.getall(x)
 
 		gui.tree(
 			DataTreeBuilder(
 				self._searchResult,
-				childrenMaker,  # lambda x: x.occurrences.items() if isinstance(x, SearchResult) else [],
+				childrenMaker,
 				labelMaker,
-				None, None, 1,
+				None,
+				None,
+				2,
 				showRoot=False,
-				onDoubleClick=lambda x: openDocument(x),
+				onDoubleClick=openDocument,
 				onContextMenu=onContextMenu
 			),
-			headerVisible=True,
-			itemDelegate=self.htmlDelegate
+			# headerVisible=True,
+			itemDelegate=gui.htmlDelegate
 		)
 
-	@property
-	def filePathsToSearch(self) -> list[FilePath]:
-		fileTypes = self._fileTypes
-		filePathsToSearch: list[FilePath] = []
-		for p in self._includedRoots:
-			if (filesIndex := p.indexBundles.get(FilesIndex)) is not None:
-				for fe in filesIndex.files.values():
-					f = fe.fullPath
-					if isinstance(f, tuple):
-						fn = f[1]
-					else:
-						fn = f
-					if not fileTypes or fn.endswith(fileTypes):
-						filePathsToSearch.append(f)
-
-		return filePathsToSearch
-
-	@TimedMethod()
-	def search(self) -> None:
-		searchResult = self._searchResult
-
+	@override
+	def prepareRun(self) -> tuple[bool, Callable[[str], Iterator[IndexSpan]]]:
+		"""
+		returns a tuple consisting of two entries:
+			- True if it is OK to run, otherwise return false
+			- an arbitrary value, which is passed on to each processFile(...) method.
+		"""
 		try:
-			try:
-				searcher = makeTextSearcher(self.searchExpr, self.searchOptions)
-			except Exception as e:
-				searchResult.error = e
-				return
-			with ZipFilePool() as zipFilePool:
-				for i, filePath in enumerate(searchResult.filesToSearch):
-					self._searchResult.filesSearched = i + 1
-					if i % 100 == 0:
-						self.progressSignal.emit(i+1)
-						QApplication.processEvents(QEventLoop.ExcludeUserInputEvents, 1)
-					try:
-						text = loadTextFile(filePath, zipFilePool)
-					except UnicodeDecodeError:
-						continue
-					lastLineStart = 0
-					lastLineNr = 0
-					for matchStart, matchEnd in searcher(text):
-						lineStart = text.rfind('\n', 0, matchStart)
-						lineStart = lineStart + 1  # skip \n at beginning of line  # if start != -1 else 0
-						end = text.find('\n', matchEnd)
-						end = end if end != -1 else len(text)
-						occurrenceStr = f'<font>{escapeForXml(text[lineStart:matchStart])}<b>{escapeForXml(text[matchStart:matchEnd])}</b>{escapeForXml(text[matchEnd:end])}</font>'
+			return True, makeTextSearcher(self.searchExpr, self.searchOptions)
+		except Exception as e:
+			self._searchResult.error = e
+			return False, lambda x: iter(())
 
-						lastLineNr += text.count('\n', lastLineStart, lineStart)
-						lastLineStart = lineStart
+	@override
+	def finishedRun(self, searcher: Callable[[str], Iterator[IndexSpan]]) -> None:
+		pass
 
-						p1 = Position(lastLineNr, matchStart - lineStart, matchStart)
-						p2 = Position(lastLineNr, matchEnd - lineStart, matchEnd)
+	@override
+	def processFile(self, filePath: FilePathTpl, pool: ZipFilePool, searcher: Callable[[str], Iterator[IndexSpan]]):
+		try:
+			text = loadTextFile(filePath, pool)
+		except UnicodeDecodeError:
+			return
 
-						searchResult.occurrences.add(filePath, Occurrence(filePath, Span(p1, p2), occurrenceStr))
-		finally:
-			self._gui.redrawGUI()
+		lastLineStart = 0
+		lastLineNr = 0
+		for matchStart, matchEnd in searcher(text):
+			lineStart = text.rfind('\n', 0, matchStart)
+			lineStart = lineStart + 1  # skip \n at beginning of line  # if start != -1 else 0
+			lineEnd = text.find('\n', matchEnd)
+			lineEnd = lineEnd if lineEnd != -1 else len(text)
+			# trim overly long lines:
+			MAX_LINE_LEN = 120
+			if lineEnd - lineStart > MAX_LINE_LEN:
+				machLen = matchEnd - matchStart
+				left = (MAX_LINE_LEN - machLen) // 2  # try to put the match in the middle.
+				left = max(15, left)  # but leave at least 15 characters before the match.
+				windowStart = max(lineStart, matchStart - left)  # don't go beyond lineStart.
+				windowEnd = min(lineEnd, matchStart + MAX_LINE_LEN)  # don't go beyond lineEnd.
+				textLeft = text[windowStart:matchStart]
+				textRight = text[matchEnd:windowEnd]
+				if windowStart > lineStart:
+					textLeft = '...' + textLeft
+				if windowEnd < lineEnd:
+					textRight = textRight + '...'
+			else:
+				textLeft = text[lineStart:matchStart]
+				textRight = text[matchEnd:lineEnd]
+
+			occurrenceStr = f'<font>{escapeForXml(textLeft)}<b>{escapeForXml(text[matchStart:matchEnd])}</b>{escapeForXml(textRight)}</font>'
+
+			lastLineNr += text.count('\n', lastLineStart, lineStart)
+			lastLineStart = lineStart
+
+			p1 = Position(lastLineNr, matchStart - lineStart, matchStart)
+			p2 = Position(lastLineNr, matchEnd - lineStart, matchEnd)
+
+			self._searchResult.occurrences.add(filePath, Occurrence(filePath, Span(p1, p2), occurrenceStr))